[Core] GetPropertiesForTypeAsync
